backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Azeem Ur Rehman",email = "rehmanazeem942@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.15"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "openai (>=2.35.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "groq (>=0.30.0,<0.31.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "langchain (>=1.2.18,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-core (>=1.3.3,<2.0.0)",
    "sentence-transformers (>=5.4.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
